Hard-code datasources to make the API core independent from the web-app.
