+++ faostat-api-core/src/test/java/org/fao/faostat/core/TestFAOSTATAPICore.java
@@ -3,6 +3,7 @@ package org.fao.faostat.core;
 import com.sun.jersey.spi.spring.container.servlet.SpringServlet;
 import com.sun.jersey.test.framework.JerseyTest;
 import com.sun.jersey.test.framework.WebAppDescriptor;
+import org.fao.faostat.core.FAOSTATAPICore;
 import org.junit.Test;
 import org.springframework.web.context.ContextLoaderListener;
 import org.springframework.web.context.request.RequestContextListener;
@@ -12,19 +13,27 @@ import static org.junit.Assert.assertNotNull;
 /**
  * @author <a href="mailto:guido.barbaglia@gmail.com">Guido Barbaglia</a>
  * */
-public class TestJSONSchemaPool extends JerseyTest {
+public class TestFAOSTATAPICore extends JerseyTest {
 
-    public TestJSONSchemaPool() {
+    public TestFAOSTATAPICore() {
         super(new WebAppDescriptor.Builder("org.fao.fenix.faostat.core").contextPath("testing")
                 .contextParam("contextConfigLocation", "classpath:testApplicationContext.xml")
                 .contextListenerClass(ContextLoaderListener.class).servletClass(SpringServlet.class)
                 .requestListenerClass(RequestContextListener.class).build());
     }
 
+    FAOSTATAPICore c;
+
     @Test
     public void testGetSchema() {
-        JSONSchemaPool j = ContextLoaderListener.getCurrentWebApplicationContext().getBean(JSONSchemaPool.class);
-        assertNotNull(j.getSchema());
+        c = ContextLoaderListener.getCurrentWebApplicationContext().getBean(FAOSTATAPICore.class);
+        assertNotNull(c.getSchema());
+    }
+
+    @Test
+    public void testGetJSONSchemaPool() {
+        c = ContextLoaderListener.getCurrentWebApplicationContext().getBean(FAOSTATAPICore.class);
+        assertNotNull(c.getJsonSchemaPool());
     }
 
 }
