+++ faostat-api-web/src/main/java/org/fao/faostat/rest/V10Classifications.java
@@ -5,6 +5,7 @@ import org.fao.faostat.beans.DatasourceBean;
 import org.fao.faostat.beans.DefaultOptionsBean;
 import org.fao.faostat.core.FAOSTATAPICore;
 import org.fao.faostat.core.StreamBuilder;
+import org.fao.faostat.datasources.DATASOURCE;
 import org.springframework.stereotype.Component;
 
 import javax.ws.rs.*;
@@ -52,7 +53,7 @@ public class V10Classifications {
             StreamBuilder sb = new StreamBuilder();
 
             /* Datasource bean. */
-            DatasourceBean datasourceBean = this.faostatapiCore.getDatasourcePool().getDatasource(this.getO().getDatasource().toUpperCase());
+            DatasourceBean datasourceBean = new DatasourceBean(DATASOURCE.valueOf(this.getO().getDatasource().toUpperCase()));
 
             /* Query the DB and create an output stream. */
             StreamingOutput stream = sb.createOutputStream("classifications", datasourceBean, this.getO());
