--- faostat-api-web/src/main/webapp/json/datasources.json
@@ -1,18 +1,0 @@
-[
-  {
-    "id": "FAOSTAT",
-    "driver": "SQLServer2000",
-    "url": "jdbc:sqlserver://HQWPRFAOSTATDB2\\Internal;databaseName=Warehouse;",
-    "dbName": "Warehouse",
-    "username": "Warehouse",
-    "password": "w@reh0use"
-  },
-  {
-    "id": "HOME-TEST",
-    "driver": "MySQL",
-    "url": "jdbc:mysql://localhost:3306/faostat",
-    "dbName": "faostat",
-    "username": "root",
-    "password": ""
-  }
-]
